--- a/benchmarks/README.md
+++ b/benchmarks/README.md
@@ -1,0 +1,7 @@
+# Benchmarks
+
+## Run benchmarks
+
+```sh
+$ ruby benchmark.rb
+```