--- a/spec/fixtures/decorators.rb
+++ b/spec/fixtures/decorators.rb
@@ -1,13 +1,8 @@
 # frozen_string_literal: true
 
-require "action_view"
 require "digest/md5"
 
-class ApplicationDecorator < Dekorator::Base; end
-
-class UserDecorator < ApplicationDecorator
-  include ActionView::Helpers::AssetTagHelper
-
+class UserDecorator < Dekorator::Base
   decorates_association :posts
   decorates_association :comments
 
@@ -15,19 +10,19 @@ class UserDecorator < ApplicationDecorator
     [first_name, last_name].join(" ")
   end
 
-  def gravatar_image_tag
-    image_tag("https://www.gravatar.com/avatar/#{Digest::MD5.hexdigest(email)}", alt: "")
+  def gravatar_image_url
+    "https://www.gravatar.com/avatar/#{Digest::MD5.hexdigest(email)}"
   end
 end
 
-class PostDecorator < ApplicationDecorator
+class PostDecorator < Dekorator::Base
   decorates_association :user
   decorates_association :comments
 end
 
 class AdvancedPostDecorator < PostDecorator; end
 
-class CommentDecorator < ApplicationDecorator
+class CommentDecorator < Dekorator::Base
   decorates_association :user
   decorates_association :post
 end