Working on some optimizations.

`master` branch

```
=================================== #summary ===================================

------------------------------------- ips --------------------------------------

Warming up --------------------------------------
            In model   145.348k i/100ms
           Dekorator     1.916k i/100ms
       Dekorator.new     1.910k i/100ms
              Draper   837.000  i/100ms
     SimpleDelegator    32.281k i/100ms
Calculating -------------------------------------
            In model      2.389M (± 2.5%) i/s -     12.064M in   5.051961s
           Dekorator     19.444k (± 6.3%) i/s -     97.716k in   5.047410s
       Dekorator.new     18.656k (± 8.1%) i/s -     93.590k in   5.052547s
              Draper      8.344k (±17.8%) i/s -     40.176k in   5.025971s
     SimpleDelegator    368.765k (± 3.0%) i/s -      1.872M in   5.082033s

Comparison:
            In model:  2389425.6 i/s
     SimpleDelegator:   368764.9 i/s - 6.48x  slower
           Dekorator:    19444.1 i/s - 122.89x  slower
       Dekorator.new:    18656.3 i/s - 128.08x  slower
              Draper:     8344.3 i/s - 286.35x  slower

------------------------------------ memory ------------------------------------

Calculating -------------------------------------
            In model     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
           Dekorator     5.552k memsize (     0.000  retained)
                        82.000  objects (     0.000  retained)
                         3.000  strings (     0.000  retained)
       Dekorator.new     5.552k memsize (     0.000  retained)
                        82.000  objects (     0.000  retained)
                         3.000  strings (     0.000  retained)
              Draper    18.799k memsize (     6.326k retained)
                       142.000  objects (    29.000  retained)
                        38.000  strings (    26.000  retained)
     SimpleDelegator   352.000  memsize (     0.000  retained)
                         4.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)

Comparison:
            In model:          0 allocated
     SimpleDelegator:        352 allocated - Infx more
           Dekorator:       5552 allocated - Infx more
       Dekorator.new:       5552 allocated - Infx more
              Draper:      18799 allocated - Infx more

================================== #comments ===================================

------------------------------------- ips --------------------------------------

Warming up --------------------------------------
            In model    96.710k i/100ms
           Dekorator     1.936k i/100ms
       Dekorator.new     1.938k i/100ms
              Draper   405.000  i/100ms
     SimpleDelegator    46.109k i/100ms
Calculating -------------------------------------
            In model      1.279M (± 3.5%) i/s -      6.480M in   5.071904s
           Dekorator     19.720k (± 7.0%) i/s -     98.736k in   5.033815s
       Dekorator.new     19.637k (± 7.3%) i/s -     98.838k in   5.062768s
              Draper      4.221k (±17.7%) i/s -     20.250k in   5.018243s
     SimpleDelegator    549.468k (± 3.8%) i/s -      2.767M in   5.042634s

Comparison:
            In model:  1279210.0 i/s
     SimpleDelegator:   549468.4 i/s - 2.33x  slower
           Dekorator:    19719.7 i/s - 64.87x  slower
       Dekorator.new:    19636.8 i/s - 65.14x  slower
              Draper:     4221.2 i/s - 303.05x  slower

------------------------------------ memory ------------------------------------

Calculating -------------------------------------
            In model   232.000  memsize (   232.000  retained)
                         1.000  objects (     1.000  retained)
                         0.000  strings (     0.000  retained)
           Dekorator     6.016k memsize (     3.192k retained)
                        84.000  objects (    23.000  retained)
                         2.000  strings (     0.000  retained)
       Dekorator.new     6.016k memsize (     3.192k retained)
                        84.000  objects (    23.000  retained)
                         2.000  strings (     0.000  retained)
              Draper    33.174k memsize (     8.126k retained)
                       238.000  objects (    48.000  retained)
                        50.000  strings (    26.000  retained)
     SimpleDelegator   352.000  memsize (   232.000  retained)
                         4.000  objects (     1.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
            In model:        232 allocated
     SimpleDelegator:        352 allocated - 1.52x more
           Dekorator:       6016 allocated - 25.93x more
       Dekorator.new:       6016 allocated - 25.93x more
              Draper:      33174 allocated - 142.99x more
```

On this branch

```
=================================== #summary ===================================

------------------------------------- ips --------------------------------------

Warming up --------------------------------------
            In model   141.224k i/100ms
           Dekorator     9.512k i/100ms
       Dekorator.new     9.734k i/100ms
              Draper   826.000  i/100ms
     SimpleDelegator    31.179k i/100ms
Calculating -------------------------------------
            In model      2.432M (± 5.2%) i/s -     12.145M in   5.008635s
           Dekorator    109.765k (± 4.5%) i/s -    551.696k in   5.036420s
       Dekorator.new    110.134k (± 4.0%) i/s -    554.838k in   5.046368s
              Draper      8.716k (±18.2%) i/s -     42.126k in   5.065191s
     SimpleDelegator    365.731k (± 3.8%) i/s -      1.840M in   5.037357s

Comparison:
            In model:  2432179.7 i/s
     SimpleDelegator:   365731.4 i/s - 6.65x  slower
       Dekorator.new:   110133.7 i/s - 22.08x  slower
           Dekorator:   109765.5 i/s - 22.16x  slower
              Draper:     8716.0 i/s - 279.05x  slower

------------------------------------ memory ------------------------------------

Calculating -------------------------------------
            In model     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
           Dekorator     3.352k memsize (     0.000  retained)
                        27.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)
       Dekorator.new     3.352k memsize (     0.000  retained)
                        27.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)
              Draper    18.799k memsize (     6.326k retained)
                       142.000  objects (    29.000  retained)
                        38.000  strings (    26.000  retained)
     SimpleDelegator   352.000  memsize (     0.000  retained)
                         4.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)

Comparison:
            In model:          0 allocated
     SimpleDelegator:        352 allocated - Infx more
           Dekorator:       3352 allocated - Infx more
       Dekorator.new:       3352 allocated - Infx more
              Draper:      18799 allocated - Infx more

================================== #comments ===================================

------------------------------------- ips --------------------------------------

Warming up --------------------------------------
            In model    97.866k i/100ms
           Dekorator   416.000  i/100ms
       Dekorator.new   430.000  i/100ms
              Draper   401.000  i/100ms
     SimpleDelegator    45.871k i/100ms
Calculating -------------------------------------
            In model      1.275M (± 4.1%) i/s -      6.361M in   4.999873s
           Dekorator      4.420k (± 8.7%) i/s -     22.048k in   5.037526s
       Dekorator.new      4.277k (± 9.4%) i/s -     21.500k in   5.080669s
              Draper      4.212k (±18.4%) i/s -     20.451k in   5.094629s
     SimpleDelegator    559.726k (± 3.7%) i/s -      2.798M in   5.006230s

Comparison:
            In model:  1274693.8 i/s
     SimpleDelegator:   559725.6 i/s - 2.28x  slower
           Dekorator:     4420.0 i/s - 288.39x  slower
       Dekorator.new:     4276.6 i/s - 298.06x  slower
              Draper:     4212.0 i/s - 302.64x  slower

------------------------------------ memory ------------------------------------

Calculating -------------------------------------
            In model   232.000  memsize (   232.000  retained)
                         1.000  objects (     1.000  retained)
                         0.000  strings (     0.000  retained)
           Dekorator    12.400k memsize (     3.192k retained)
                       145.000  objects (    23.000  retained)
                         7.000  strings (     0.000  retained)
       Dekorator.new    12.400k memsize (     3.192k retained)
                       145.000  objects (    23.000  retained)
                         7.000  strings (     0.000  retained)
              Draper    33.174k memsize (     8.126k retained)
                       238.000  objects (    48.000  retained)
                        50.000  strings (    26.000  retained)
     SimpleDelegator   352.000  memsize (   232.000  retained)
                         4.000  objects (     1.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
            In model:        232 allocated
     SimpleDelegator:        352 allocated - 1.52x more
           Dekorator:      12400 allocated - 53.45x more
       Dekorator.new:      12400 allocated - 53.45x more
              Draper:      33174 allocated - 142.99x more
```