@nicolas-brousse 2019-11-27
Not sure about the credibility of the benchmark. I'm not use to write benchmarks :/

I took some improvements I wrote in #9 since it's better to not use that mush methods like `#blank?`, `#prensent?` and `decorate`.

I think is still a good improvement for v1.